BREAKING: Remove `skip-replies`
This was a half-baked and ill-considered feature that interacts
poorly with manual pipeline control (`with-reply`, etc.) and doesn't
serve any important purpose since the addition of manual pipeline
control.

As an added benefit, its removal also speeds reply processing.

This is a breaking change, but hopefully won't affect a large number
of people since it was a pretty obscure feature anyway and was mostly
used internally.