Parsers no longer mask exceptions by default
This is much more sensible behaviour and fixes a bug when `evalsha*` was
enclosed in a `with-parser` context. As a bonus, the new behaviour improves
general reply-parsing performance.

To reproduce the old behaviour (e.g. to write a parser that deals manually
with exceptions), attach {:throw-exceptions? false} metadata to your parser
fn.