WIP/nb: removed multi-based CAS utils
Benchmarking showed a small->negligible perf difference between
the 2nd-gen Lua-based and multi-based cases for both small and
large old-vals.

The Lua-based approach has the additional advantages of being
more flexible, and *much* faster in the high-contention hash
case (which may be quite common).

Offering both pairs of approaches seems excessive, so going to
double down on the Lua.