Message queue: Block for worker stop
When running the `mq/stop`, deref all the worker-loop futures.

This is helpful if, for example, the worker processing depends on other
parts of a component system. At the moment I'm getting errors where the
component system shuts down without being aware of the fact that there
are still some worker threads running, meaning that the workers try to
access closed resources (for example).