Correct misleading readme
> "Which means that not only is Carmine's command coverage *always complete*,..."

This seems not true as far as I can tell. For example, the official Redis command reference has [`LMOVE`](https://github.com/redis/redis-doc/blob/master/commands.json#L7744) (introduced in v6.2.0) but this is not in the current version 3.0.0 of Carmine.

I appreciate there's an effort in keeping Carmine up to date with the latest and greatest Redis has to offer but saying it's "*always complete*" is misleading.